some renames and reduced scope of global imports
